+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [0.0.0] - 04/11/2025 
+
+### Added
+- Merge with realese 0-1-0
+- Update supplies model
+
+## [0.0.0] - 02/11/2025 
+
+### Added
+- Updated supplies model to follow database names
+- Implemented routes, controller, and model for the Get Supplies List use case.
+
 ## [0.1.0] - 01/11/2025  
 
 ### Added